fix(spindrift): cut the client's streaming value-edge into db/
Ticket 34, edge 2 (edge 1 and the guard landed in #1496). app.tsx ->
stream-client.ts value-imported the stream paths and message types
straight from streams.ts, which value-imports drizzle-orm, db/notify.ts,
and db/schema.ts to serve the WebSocket transport — the same
resolve-before-tree-shake landmine PR #1492's CI hit on a file it never
touched.

Moves ATTEMPT_STREAM_PATH, RUNTIME_STREAM_PATH, STREAM_PATHS, and the
stream message types into src/web/stream-path.ts, taking
AttemptLogCursor/AttemptLogEntry and RuntimeLogPage as `import type`
only so Bun never turns them into a module edge — the same pattern
command-path.ts established for edge 1. streams.ts re-exports from
there rather than duplicating the definitions, so the server side is
unchanged.

Extends test/web/client-bundle.test.ts with a second assertion that
web/streams.ts and db/notify.ts are absent from the built bundle.
